requirements.txt
boto3==1.26.78
botocore==1.29.78
datefinder==0.7.3
jmespath==1.0.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
regex==2022.10.31
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
